requirements.txt
canvasapi==3.2.0
py-trello==0.19.0
parse==1.19.0
pyinstaller==5.12.0
python-dotenv==0.21.1